+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.3.0
 
+- Add support for user review lists (`/screeny05/reviews/`) (Thank you @benjroy)
 - Add support for Studio filmographies (`/studio/a24/`, etc.)
 - Only the default (most popular) page is returned. Further pagination is blocked by Cloudflare (see `lib/letterboxd/studio.ts`).
 - Fix broken write-after-flush in chunkstreamer